refactor: add data directory and locate repository and storage directory
