+++ src/data/repository/SettingsRepository.ts
@@ -1,6 +1,5 @@
-import { PopupSize, defaultPopupSize } from "../model/PopupSize";
-
-import { ChromeLocalStorage } from "./ChromeStorage";
+import { PopupSize, defaultPopupSize } from "../../model/PopupSize";
+import { ChromeLocalStorage } from "../storage/ChromeStorage";
 
 export const getPopupSizeSetting = async (): Promise<PopupSize> => {
   const setting = await ChromeLocalStorage.getPopupSizeSetting();
