+++ src/data/repository/TabGroupRepository.ts
@@ -1,7 +1,10 @@
-import { StoredTabGroup, TabGroup, TabGroupColor } from "../model/TabContainer";
-import { WindowId } from "../model/Window";
-
-import { ChromeLocalStorage } from "./ChromeStorage";
+import {
+  StoredTabGroup,
+  TabGroup,
+  TabGroupColor,
+} from "../../model/TabContainer";
+import { WindowId } from "../../model/Window";
+import { ChromeLocalStorage } from "../storage/ChromeStorage";
 
 export const collapseTabGroup = async (groupId: number): Promise<void> => {
   await chrome.tabGroups.update(groupId, { collapsed: true });
