+++ src/view/components/OrganizationPage/TabGroupingForm.tsx
@@ -10,15 +10,15 @@ import MenuItem from "@mui/material/MenuItem";
 import Select from "@mui/material/Select";
 import Switch from "@mui/material/Switch";
 import { useEffect, useState } from "react";
-import t from "../../../i18n/Translations";
-import { TabGroupSetting } from "../../../model/TabGroupSetting";
 import {
   addListenerOnUpdateTabGroupSetting,
   getTabGroupSetting,
   groupTabsBySetting,
   removeListenerOnUpdateTabGroupSetting,
   updateTabGroupSetting,
-} from "../../../repository/TabGroupSettingRepository";
+} from "../../../data/repository/TabGroupSettingRepository";
+import t from "../../../i18n/Translations";
+import { TabGroupSetting } from "../../../model/TabGroupSetting";
 import PaperWithHeader from "../PaperWithHeader";
 
 type TabGroupingFormProps = {
