+++ src/view/components/RestorePage/StoredTabGroups.tsx
@@ -17,9 +17,9 @@ import Grid from "@mui/material/Unstable_Grid2";
 import { grey } from "@mui/material/colors";
 import { alpha, styled, useTheme } from "@mui/material/styles";
 import { useContext, useEffect, useRef, useState } from "react";
+import { restoreTabGroup } from "../../../data/repository/TabGroupRepository";
 import t from "../../../i18n/Translations";
 import { StoredTabGroup, tabGroupColors } from "../../../model/TabContainer";
-import { restoreTabGroup } from "../../../repository/TabGroupRepository";
 import { StoredTabGroupsContext } from "../../contexts/StoredTabGroupsContext";
 import { useRemoveStoredTabGroup } from "../../hooks/useRemoveStoredTabGroup";
 import { useUpdateStoredTabGroupColor } from "../../hooks/useUpdateStoredTabGroupColor";
