+++ src/view/components/RestorePage/StoredWindows.tsx
@@ -17,9 +17,9 @@ import Grid from "@mui/material/Unstable_Grid2";
 import { grey } from "@mui/material/colors";
 import { alpha, styled, useTheme } from "@mui/material/styles";
 import { useContext, useEffect, useRef, useState } from "react";
+import { restoreWindow } from "../../../data/repository/WindowsRepository";
 import t from "../../../i18n/Translations";
 import { StoredWindow } from "../../../model/Window";
-import { restoreWindow } from "../../../repository/WindowsRepository";
 import { StoredWindowsContext } from "../../contexts/StoredWindowsContext";
 import { useRemoveStoredWindow } from "../../hooks/useRemoveStoredWindow";
 import { useUpdateStoredWindowName } from "../../hooks/useUpdateStoredWindowName";
