+++ src/view/components/TabItem.tsx
@@ -14,10 +14,10 @@ import Typography from "@mui/material/Typography";
 import grey from "@mui/material/colors/grey";
 import { useTheme } from "@mui/material/styles";
 import { forwardRef, useContext, useState } from "react";
+import { closeTab, focusTab } from "../../data/repository/TabsRepository";
 import t from "../../i18n/Translations";
 import { Tab, durationSinceLastActivatedAt } from "../../model/Tab";
 import { findTabGroup, hasDuplicatedTabs } from "../../model/Window";
-import { closeTab, focusTab } from "../../repository/TabsRepository";
 import { WindowsContext } from "../contexts/WindowsContext";
 import { resolveDuplicatedTabs } from "../functions/resolveDuplicatedTabs";
 import { TabItemActionMenu } from "./ActionMenu";
