+++ src/view/features/popup/components/Header.tsx
@@ -15,8 +15,8 @@ import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
 import { alpha, styled } from "@mui/material/styles";
 import { useContext, useState } from "react";
+import { navigateToOptionsPage } from "../../../../data/repository/SettingsRepository";
 import t from "../../../../i18n/Translations";
-import { navigateToOptionsPage } from "../../../../repository/SettingsRepository";
 import { PopupHeaderActionMenu } from "../../../components/ActionMenu";
 import { ModeContext } from "../../../contexts/ModeContext";
 import { useToggleMode } from "../../../hooks/useToggleMode";
