+++ src/view/hooks/useSaveStoredTabGroup.ts
@@ -1,7 +1,7 @@
 import { useCallback, useContext } from "react";
 
+import { saveStoredTabGroup } from "../../data/repository/TabGroupRepository";
 import { TabGroup } from "../../model/TabContainer";
-import { saveStoredTabGroup } from "../../repository/TabGroupRepository";
 import { StoredTabGroupsContext } from "../contexts/StoredTabGroupsContext";
 
 export const useSaveStoredTabGroup = (): ((
