+++ src/view/hooks/useSaveStoredWindow.ts
@@ -1,7 +1,7 @@
 import { useCallback, useContext } from "react";
 
+import { saveStoredWindow } from "../../data/repository/WindowsRepository";
 import { Window } from "../../model/Window";
-import { saveStoredWindow } from "../../repository/WindowsRepository";
 import { StoredWindowsContext } from "../contexts/StoredWindowsContext";
 
 export const useSaveStoredWindow = (): ((window: Window) => Promise<void>) => {
